course-project-group32-main/requirements.txt
Flask
numpy
soundfile
ffmpeg-python
moviepy
Werkzeug
PyJWT
mysql-connector-python
pillow
psycopg2-binary
pydub
opencv-python
matplotlib
librosa
webcolors
